[IMPROVE] Modularize project into packages. (v1.15)
