+++ python_easy_chess_gui/__init__.py
@@ -1,0 +1,4 @@
+from pathlib import Path
+
+
+ROOT_PATH = Path(__file__).parent.resolve()
