+++ python_easy_chess_gui/play_package/play_module.py
@@ -1,0 +1,108 @@
+import logging
+import os
+
+import chess
+
+
+class Timer:
+    def __init__(self, tc_type='fischer', base=300000, inc=10000,
+                 period_moves=40):
+        """
+        :param tc_type: time control type ['fischer, delay, classical']
+        :param base: base time in ms
+        :param inc: increment time in ms can be negative and 0
+        :param period_moves: number of moves in a period
+        """
+        self.tc_type = tc_type  # ['fischer', 'delay', 'timepermove']
+        self.base = base
+        self.inc = inc
+        self.period_moves = period_moves
+        self.elapse = 0
+        self.init_base_time = self.base
+
+    def update_base(self):
+        """
+        Update base time after every move
+
+        :return:
+        """
+        if self.tc_type == 'delay':
+            self.base += min(0, self.inc - self.elapse)
+        elif self.tc_type == 'fischer':
+            self.base += self.inc - self.elapse
+        elif self.tc_type == 'timepermove':
+            self.base = self.init_base_time
+        else:
+            self.base -= self.elapse
+
+        self.base = max(0, self.base)
+        self.elapse = 0
+
+
+class GuiBook:
+    def __init__(self, book_file, board, is_random=True):
+        """
+        Handle gui polyglot book for engine opponent.
+
+        :param book_file: polgylot book filename
+        :param board: given board position
+        :param is_random: randomly select move from book
+        """
+        self.book_file = book_file
+        self.board = board
+        self.is_random = is_random
+        self.__book_move = None
+
+    def get_book_move(self):
+        """ Returns book move either random or best move """
+        reader = chess.polyglot.open_reader(self.book_file)
+        try:
+            if self.is_random:
+                entry = reader.weighted_choice(self.board)
+            else:
+                entry = reader.find(self.board)
+            self.__book_move = entry.move
+        except IndexError:
+            logging.warning('No more book move.')
+        except Exception:
+            logging.exception('Failed to get book move.')
+        finally:
+            reader.close()
+
+        return self.__book_move
+
+    def get_all_moves(self):
+        """
+        Read polyglot book and get all legal moves from a given positions.
+
+        :return: move string
+        """
+        is_found = False
+        total_score = 0
+        book_data = {}
+        cnt = 0
+
+        if os.path.isfile(self.book_file):
+            moves = '{:4s}   {:<5s}   {}\n'.format('move', 'score', 'weight')
+            with chess.polyglot.open_reader(self.book_file) as reader:
+                for entry in reader.find_all(self.board):
+                    is_found = True
+                    san_move = self.board.san(entry.move)
+                    score = entry.weight
+                    total_score += score
+                    bd = {cnt: {'move': san_move, 'score': score}}
+                    book_data.update(bd)
+                    cnt += 1
+        else:
+            moves = '{:4s}  {:<}\n'.format('move', 'score')
+
+        # Get weight for each move
+        if is_found:
+            for _, v in book_data.items():
+                move = v['move']
+                score = v['score']
+                weight = score/total_score
+                moves += '{:4s}   {:<5d}   {:<2.1f}%\n'.format(move, score,
+                                                            100*weight)
+
+        return moves, is_found
